psens/requirements.txt
argparse
paho-mqtt
pip
setuptools
wsgiref
pyowm
